# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gigs API",
        "description": "API documentation for Gigs API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company/{id}/gig": {
            "get": {
                "tags": [
                    "Company/Gig"
                ],
                "summary": "GET (all) gigs",
                "description": "Returns list of gigs",
                "operationId": "GetGigs",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GigResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company/Gig"
                ],
                "summary": "POST (add) gig",
                "description": "Returns gig",
                "operationId": "addGig",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GigRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GigResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/{company_id}/gig/{gig_id}": {
            "get": {
                "tags": [
                    "Company/Gig"
                ],
                "summary": "GET Gig",
                "description": "Show Gig",
                "operationId": "showGig",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "gig_id",
                        "in": "path",
                        "description": "Gig id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GigResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Company/Gig"
                ],
                "summary": "UPDATE Gig",
                "description": "UPDATE gig",
                "operationId": "updateGig",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "gig_id",
                        "in": "path",
                        "description": "Gig id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GigUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GigResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company/Gig"
                ],
                "summary": "DELETE Gig",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteGig",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "gig_id",
                        "in": "path",
                        "description": "Gig id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource was deleted successfully."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/gigs/search/{term}": {
            "get": {
                "tags": [
                    "Gig/Search"
                ],
                "summary": "GET SearchGigs",
                "description": "Search gigs by term (both name and description)",
                "operationId": "searchGigs",
                "parameters": [
                    {
                        "name": "term",
                        "in": "path",
                        "description": "Term",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful searching",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GigResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/gigs/filter": {
            "get": {
                "tags": [
                    "Gig/Filter"
                ],
                "summary": "GET FilterGigs",
                "description": "Filter gigs by company_id, progress(not_started, started, finished) and status(Draft, Posted)",
                "operationId": "filterGigs",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Filter by company ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "progress",
                        "in": "query",
                        "description": "Filter by progress(not_started, started, finished)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "not_started",
                                "started",
                                "finished"
                            ],
                            "example": "started"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status(0 - Draft, 1 -Posted)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful filtering",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GigResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GigRequest": {
                "title": "Create Gig Request",
                "description": "Create Gig request body data",
                "required": [
                    "name",
                    "timestamp_start",
                    "timestamp_end",
                    "number_of_positions",
                    "pay_per_hour"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Gig's name",
                        "type": "string",
                        "example": "Gig name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the Gig",
                        "type": "string",
                        "example": "Doe"
                    },
                    "timestamp_start": {
                        "title": "Timestamp start",
                        "description": "Gig's timestamp start",
                        "type": "string",
                        "format": "date-time",
                        "example": "10/25/2024 12:00 AM"
                    },
                    "timestamp_end": {
                        "title": "Timestamp end",
                        "description": "Gig's timestamp end",
                        "type": "string",
                        "format": "date-time",
                        "example": "10/27/2024 12:00 PM"
                    },
                    "number_of_positions": {
                        "title": "Number of positions",
                        "description": "Gig's number of positions",
                        "type": "integer",
                        "example": "10"
                    },
                    "pay_per_hour": {
                        "title": "Pay per hour",
                        "description": "Gig's pay per hour",
                        "type": "number",
                        "format": "double",
                        "example": 35.5
                    },
                    "status": {
                        "title": "Status",
                        "description": "Gig's status",
                        "type": "boolean",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "GigUpdateRequest": {
                "title": "Update Gig Request",
                "description": "Update Gig request body data",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Gig's name",
                        "type": "string",
                        "example": "Gig name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the Gig",
                        "type": "string",
                        "example": "Doe"
                    },
                    "timestamp_start": {
                        "title": "Timestamp start",
                        "description": "Gig's timestamp start",
                        "type": "string",
                        "format": "date-time",
                        "example": "10/25/2024 12:00 AM"
                    },
                    "timestamp_end": {
                        "title": "Timestamp end",
                        "description": "Gig's timestamp end",
                        "type": "string",
                        "format": "date-time",
                        "example": "10/27/2024 12:00 PM"
                    },
                    "number_of_positions": {
                        "title": "Number of positions",
                        "description": "Gig's number of positions",
                        "type": "integer",
                        "example": "10"
                    },
                    "pay_per_hour": {
                        "title": "Pay per hour",
                        "description": "Gig's pay per hour",
                        "type": "number",
                        "format": "double",
                        "example": 35.5
                    },
                    "status": {
                        "title": "Gig's status",
                        "description": "Gig's status",
                        "type": "boolean",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Gig": {
                "title": "Gig",
                "description": "Gig model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Gig's name",
                        "type": "string",
                        "example": "Gig name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the Gig",
                        "type": "string",
                        "example": "Doe"
                    },
                    "timestamp_start": {
                        "title": "Timestamp start",
                        "description": "Gig's timestamp start",
                        "type": "string",
                        "format": "date-time",
                        "example": "10/25/2024 12:00 AM"
                    },
                    "timestamp_end": {
                        "title": "Timestamp end",
                        "description": "Gig's timestamp end",
                        "type": "string",
                        "format": "date-time",
                        "example": "10/27/2024 12:00 PM"
                    },
                    "number_of_positions": {
                        "title": "Number of positions",
                        "description": "Gig's number of positions",
                        "type": "integer",
                        "example": "10"
                    },
                    "pay_per_hour": {
                        "title": "Pay per hour",
                        "description": "Gig's pay per hour",
                        "type": "number",
                        "format": "double",
                        "example": 35.5
                    },
                    "status": {
                        "title": "Status",
                        "description": "Gig's status",
                        "type": "boolean",
                        "example": 0
                    },
                    "company": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "name": {
                                "type": "string",
                                "example": "Company name"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Gig"
                }
            },
...
            "GigResource": {
                "title": "GigResource",
                "description": "Gig resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Gig"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GigResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://gigs.devs/oauth/authorize",
                        "tokenUrl": "http://gigs.devs/oauth/token",
                        "refreshUrl": "http://gigs.devs/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}